+++ tests/snapshot_tests/test_snapshots.py
@@ -4626,3 +4626,67 @@ def test_header_format(snap_compare):
             yield Header()
 
     assert snap_compare(HeaderApp())
+
+
+def test_long_textarea_placeholder(snap_compare) -> None:
+    """Test multi-line placeholders are wrapped and rendered.
+    You should see a TextArea at 50% width, with several lines of wrapped text.
+    """
+
+    TEXT = """I must not fear.
+Fear is the mind-killer.
+Fear is the little-death that brings total obliteration.
+I will face my fear.
+I will permit it to pass over me and through me.
+And when it has gone past, I will turn the inner eye to see its path.
+Where the fear has gone there will be nothing. Only I will remain."""
+
+    class PlaceholderApp(App):
+
+        CSS = """
+        TextArea {
+            width: 50%;
+        }
+        """
+
+        def compose(self) -> ComposeResult:
+            yield TextArea(placeholder=TEXT)
+
+    assert snap_compare(PlaceholderApp())
+
+
+def test_rich_log_early_write(snap_compare) -> None:
+    """Regression test for https://github.com/Textualize/textual/issues/6123
+
+    You should see a RichLog with "Hello World" text
+
+    """
+
+    class TestApp(App):
+        def compose(self) -> ComposeResult:
+            with Horizontal():
+                yield RichLog()
+
+        def on_mount(self) -> None:
+            self.theme = "nord"
+
+        def on_ready(self) -> None:
+            log_widget = self.query_one(RichLog)
+            log_widget.write("Hello, World!")
+
+    assert snap_compare(TestApp())
+
+
+def test_collapsible_focus_children(snap_compare) -> None:
+    """Regression test for https://github.com/Textualize/textual/issues/6140
+
+    You should see an expanded collapsible containing a button. The button should be focused.
+
+    """
+
+    class CollapseApp(App):
+        def compose(self) -> ComposeResult:
+            with Collapsible(title="Collapsible", collapsed=False):
+                yield Button("Hello")
+
+    assert snap_compare(CollapseApp(), press=["enter", "enter", "tab"])